requirements.txt
streamlit>=1.37.0
requests>=2.32.0
python-dateutil>=2.9.0
pandas>=2.2.0



